Upgrade SceneManager to use etna and purge a lot of vk_utils
SceneManager now uses etna to deal with buffers, and so we can finally remove, like, most of vk_utils.

The only remaining legacy part is mesh management. For that, I want to migrate to glTF (which isn't good, but is better than ad-hoc formats and is good enough for our purposes).
Afterwards, as much stuff as possible should be repackaged into etna and we are golden.